# Conflicts:
#	.github/workflows/lint.yml
#	tests/integration/call_cases.py
#	tests/integration/language_specs.py
#	tests/integration/test_literalize_golden.py

+++ tests/integration/cases/scalar_int_large/V_integer_format_hex.v
@@ -1,5 +1,5 @@
 
 fn main() {
-	my_data := 0x80000000
+	my_data := i64(0x80000000)
 	_ = my_data
 }
